feature/blocs (#27)
* Blocs por todo lado

* Login

* Bloc on appointments

* Bloc on professions

* Change variable name

* All providers

* Bloc on professionals

* Bloc on professions

* Bloc on reviews

* Bloc on specialties

* Bloc on summary

* Pesquisa profissional por especialidade

* Prevent fail when summary data is null

+++ lib/models/feeling.dart
@@ -6,7 +6,7 @@ class Feeling {
 
   Feeling.fromJson(Map<String, dynamic> json) {
     count = json['count'];
-    percent = (json['percent'] as int).toDouble();
+    percent = (json['percent'] as double).toDouble();
   }
 
   Map<String, dynamic> toJson() {
